requirements.txt
Django==5.0.1
djangorestframework==3.14.0
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1
